fix(e2e): address fork review feedback
