+++ wework/e2e/desktop/task-flow.e2e.mjs
@@ -393,6 +393,44 @@ function validateDesktopSegmentOptions() {
   if (DESKTOP_SEGMENT && DESKTOP_FROM_SEGMENT) {
     throw new Error('--segment and --from-segment cannot be used together')
   }
+  const activeOnlyModes = [
+    ['WEWORK_DESKTOP_E2E_REQUEST_INPUT_ONLY=1', REQUEST_INPUT_ONLY],
+    ['--view-image-only', VIEW_IMAGE_ONLY],
+    ['--short-conversation-only', SHORT_CONVERSATION_ONLY],
+    ['--retry-only', RETRY_ONLY],
+    ['--rate-limit-only', RATE_LIMIT_ONLY],
+    ['--running-fork-only', RUNNING_FORK_ONLY],
+    ['--completed-fork-only', COMPLETED_FORK_ONLY],
+    ['--side-chat-only', SIDE_CHAT_ONLY],
+    ['--goal-idle-only', GOAL_IDLE_ONLY],
+    ['--goal-restart-only', GOAL_RESTART_ONLY],
+    ['--turn-navigation-only', TURN_NAVIGATION_ONLY],
+    ['--attachment-only', ATTACHMENT_ONLY],
+    ['--pasted-workspace-paths-only', PASTED_WORKSPACE_PATHS_ONLY],
+    ['--dropped-workspace-paths-only', DROPPED_WORKSPACE_PATHS_ONLY],
+    ['--system-drag-panel-only', SYSTEM_DRAG_PANEL_ONLY],
+    ['--model-switch-only', MODEL_SWITCH_ONLY],
+    ['--cloud-only', CLOUD_ONLY],
+    ['--plugins-only', PLUGINS_ONLY],
+    ['--automation-only', AUTOMATION_ONLY],
+    ['--memory-only', MEMORY_ONLY],
+    ['--tool-block-order-only', TOOL_BLOCK_ORDER_ONLY],
+    ['--queue-navigation-only', QUEUE_NAVIGATION_ONLY],
+    ['--guidance-background-only', GUIDANCE_BACKGROUND_ONLY],
+    ['--guidance-scroll-only', GUIDANCE_SCROLL_ONLY],
+    ['--message-restoration-only', MESSAGE_RESTORATION_ONLY],
+    ['--queue-management-only', QUEUE_MANAGEMENT_ONLY],
+    ['--task-plan-only', TASK_PLAN_ONLY],
+    ['WEWORK_E2E_DESKTOP_SCENARIO_ONLY=true', DESKTOP_SCENARIO_ONLY],
+    ['WEWORK_E2E_MIXED_TOOL_TURNS_ONLY=1', MIXED_TOOL_TURNS_ONLY],
+  ].filter(([, enabled]) => enabled)
+  if (activeOnlyModes.length > 1) {
+    throw new Error(
+      `Desktop E2E only modes are mutually exclusive: ${activeOnlyModes
+        .map(([name]) => name)
+        .join(', ')}`
+    )
+  }
   const availableSegments = [...DESKTOP_CHECKPOINTS, ...PLUGIN_SEGMENTS]
   if (SELECTED_DESKTOP_SEGMENT && !availableSegments.includes(SELECTED_DESKTOP_SEGMENT)) {
     throw new Error(
@@ -2094,10 +2132,6 @@ async function verifyCompletedTurnFork({
     FORK_FOLLOW_UP_PROMPT,
     'fork_follow_up'
   )
-  await control.command('waitFor', '[data-testid="message-assistant"]', {
-    text: FORK_FOLLOW_UP_COMPLETION_TEXT,
-    timeoutMs: DEFAULT_STEP_TIMEOUT_MS,
-  })
   assert.ok(
     JSON.stringify(forkFollowUpRequest.body).includes(FORK_FOLLOW_UP_PROMPT),
     'The forked task did not accept an independent follow-up'
@@ -2107,6 +2141,10 @@ async function verifyCompletedTurnFork({
     false,
     'The forked request forwarded opaque encrypted reasoning history'
   )
+  await control.command('waitFor', '[data-testid="message-assistant"]', {
+    text: FORK_FOLLOW_UP_COMPLETION_TEXT,
+    timeoutMs: DEFAULT_STEP_TIMEOUT_MS,
+  })
   await captureVerificationScreenshot(control, 'completed-turn-fork-03-follow-up-complete.png')
 
   await control.command('click', `[data-testid="${sourceTaskRowTestId}"]`)
@@ -10458,6 +10496,23 @@ async function main() {
         return
       }
 
+      if (RUNNING_FORK_ONLY) {
+        phase = 'running-follow-up-fork'
+        await verifyRunningFollowUpFork({
+          composerSelector,
+          control,
+          executorHome,
+          sourceTaskRowTestId: taskRowTestId,
+        })
+        await writeFile(
+          join(resultDir, 'model-requests.json'),
+          `${JSON.stringify(control.modelRequests, null, 2)}\n`,
+          'utf8'
+        )
+        console.log(`Wework running-fork desktop E2E passed. Evidence: ${resultDir}`)
+        return
+      }
+
       if (!COMPLETED_FORK_ONLY) {
         phase = 'running-follow-up-fork'
         await verifyRunningFollowUpFork({
@@ -10466,15 +10521,6 @@ async function main() {
           executorHome,
           sourceTaskRowTestId: taskRowTestId,
         })
-        if (RUNNING_FORK_ONLY) {
-          await writeFile(
-            join(resultDir, 'model-requests.json'),
-            `${JSON.stringify(control.modelRequests, null, 2)}\n`,
-            'utf8'
-          )
-          console.log(`Wework running-fork desktop E2E passed. Evidence: ${resultDir}`)
-          return
-        }
       }
 
       phase = 'completed-turn-fork'
